sums and getting data from file moved to in plotFile calss
[doc] runtime commands - diagram

[doc] doxygen publish script, main page

added missed png

[doc] connected RPC commands

[doc] small changes in doc

[doc] fixed links

[doc] changes from README restored; Build, Make

+++ utils/dr-monero/utils.h
@@ -21,8 +21,8 @@ class utils
 {
 public:
     utils();
-    vector<double>  simpleSmooth(vector<double> data, double alpha);
-    vector<double> prepareHistogramData(vector<double> t, vector<double> b, int frame);
+    vector<long double>  simpleSmooth(vector<long double> data, double alpha);
+    vector<long double> prepareHistogramData(vector<long double> t, vector<long double> b, int frame);
 private:
     void display(vector <double> data);
 };
